% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{adjust_seeding_ages}
\alias{adjust_seeding_ages}
\title{Adjust seeding ages}
\usage{
adjust_seeding_ages(initial_values, parameters)
}
\arguments{
\item{initial_values}{Vector of inital values from with variables object
created by \code{\link{create_variables}}}

\item{parameters}{Parameters object created by \code{\link{get_parameters}}}
}
\value{
Returns modified initial_values vector
}
\description{
Adjust seeding ages
}
\details{
Switches age variables based on parameters object
}
\examples{
\dontrun{
Create our parameters
pop <- squire::get_population(iso3c = "ATG")
pop$n <- as.integer(pop$n)/100
parameters <- get_parameters(
   population = pop$n, contact_matrix_set = squire::contact_matrices[1]
)

# Create our variables
variables <- create_variables(pop)

# adjust the seeding ages
variables$discrete_age$initial_values <- adjust_seeding_ages(
initial_values = variables$discrete_age$initial_values,
  parameters = parameters
)
}
}
