% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_stochastic.R
\name{clear_schedule}
\alias{clear_schedule}
\title{Clear schedules when individuals die}
\usage{
clear_schedule(died)
}
\arguments{
\item{died}{}
}
\value{

}
\description{
Clear schedules when individuals die
}
