% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{Create_age_variable}
\alias{Create_age_variable}
\title{Continuous age variable}
\usage{
Create_age_variable(pop)
}
\arguments{
\item{pop}{}
}
\value{
age variable
}
\description{
Create a continuous age variable for the population
}
\examples{
Create_age_variable(pop)
}
