% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_continuous_age_variable}
\alias{create_continuous_age_variable}
\title{Continuous age variable}
\usage{
create_continuous_age_variable(pop)
}
\arguments{
\item{pop}{population list}
}
\value{
continuous age variable
}
\description{
Create a continuous age variable for the population
}
