% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_exposure_update_listener}
\alias{create_exposure_update_listener}
\title{Modelling the progression to either IMild or ICase}
\usage{
create_exposure_update_listener(human, states, events, variables, parameters)
}
\arguments{
\item{human}{the handle for the human individuals}

\item{states}{the available human states}

\item{variables}{the available human variables}

\item{parameters}{model parameters}
}
\description{
Age dependent outcome of exposure
}
