% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_hospitilisation_update_listener}
\alias{create_hospitilisation_update_listener}
\title{Modelling the progression to either IMv or IOx}
\usage{
create_hospitilisation_update_listener(
  human,
  states,
  variables,
  parameters,
  events
)
}
\arguments{
\item{human}{the handle for the human individuals}

\item{states}{the available human states}

\item{variables}{the available human variables}

\item{parameters}{model parameters}
}
\description{
Age dependent outcome of exposure
}
