% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_human}
\alias{create_human}
\title{Define the human model}
\usage{
create_human(states, variables, events, parameters)
}
\arguments{
\item{states}{available states to assign}

\item{variables}{available variables to assign}

\item{events}{available events to assign}

\item{parameters}{available parameters to assign}
}
\value{
human - human
}
\description{
Declares the human individual and assigns the
relevant states and variables
}
