% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_individuals}
\alias{create_individuals}
\title{Define model individuals}
\usage{
create_individuals(states)
}
\arguments{
\item{states}{list of states}
}
\description{
Create_individuals declares the individuals to simulate. It assigns the
relevant states and variables to each individual.
}
