% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_infection_update_listener}
\alias{create_infection_update_listener}
\title{Update the state of an individual as infection events occur}
\usage{
create_infection_update_listener(human, to_state)
}
\arguments{
\item{human}{the handle for the human individuals}

\item{to_state}{the destination disease state}
}
\description{
Randomly moves individuals towards the later stages of disease
and updates their ineffectivity
}
