% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create processes for simulation}
\usage{
create_processes(individualname, statenames)
}
\arguments{
\item{individualname}{name of individual}

\item{statenames}{names of states}
}
\description{
Create processes for simulation
}
