% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{create_progression_listener}
\alias{create_progression_listener}
\title{Modelling the progression of the human disease}
\usage{
create_progression_listener(event, duration, shift = 0, func = r_erlang)
}
\arguments{
\item{event}{the event to schedule}

\item{duration}{the average time spent in this state}

\item{shift}{number of days to increase scheduled event by}

\item{func}{function to use for drawing progression.
Default = \link{r_erlang}}
}
\description{
schedules follow up infection events
}
