% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_setup_process}
\alias{create_setup_process}
\title{Define event based processes at initialisation}
\usage{
create_setup_process(individuals, states, events, variables)
}
\arguments{
\item{individuals}{a list of individuals in the model}

\item{states}{a list of states in the model}

\item{events}{a list of events in the model}

\item{variables}{list of variables in the model}
}
\description{
Defines processes for events for states at initialisation
}
