% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{Create_states}
\alias{Create_states}
\title{Define model states}
\usage{
Create_states(psq)
}
\arguments{
\item{psq}{the model parameters}
}
\value{
states
}
\description{
Create_states creates and initialises the human states for the model
}
