% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_variables}
\alias{create_variables}
\title{Define model variables}
\usage{
create_variables(psq)
}
\arguments{
\item{psq, }{model parameters created by \code{get_parameters}}
}
\description{
create_variables creates human variables
}
