% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Default_probs}
\alias{Default_probs}
\title{Return the default probabilities for modelling}
\usage{
Default_probs()
}
\value{
list of default probabilities
}
\description{
Return the default probabilities for modelling
}
