% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{Div_pop}
\alias{Div_pop}
\title{Divide matrix by population}
\usage{
Div_pop(contact, population)
}
\arguments{
\item{contact}{contact}

\item{population}{Vector of population by age}
}
\value{
Matrix
}
\description{
Divide matrix by population
}
