% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{doing_the_swap}
\alias{doing_the_swap}
\title{Swap values that have been identified}
\usage{
doing_the_swap(swaps, age)
}
\arguments{
\item{swaps}{values to be swapped}

\item{age}{discrete age or continuous age}
}
\value{
age
}
\description{
Swap values that have been identified
}
