% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{E2_IMild}
\alias{E2_IMild}
\title{Function for transition from E2 to IMild}
\usage{
E2_IMild(human, IMild, E2, ICase1, p_E2_I, prob_hosp)
}
\arguments{
\item{human}{human}

\item{IMild}{IMild}

\item{E2}{E2}

\item{ICase1}{ICase1}

\item{p_E2_I}{p_E2_I}

\item{prob_hosp}{prob_hosp}
}
\description{
Function for transition from E2 to IMild
}
