% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{E_IMild_ICase1_ICase2}
\alias{E_IMild_ICase1_ICase2}
\title{Function for transition from E to IMild and ICase1 and ICase2}
\usage{
E_IMild_ICase1_ICase2(
  human,
  IMild,
  E,
  ICase1,
  E_I,
  p_E2_I,
  prob_hosp,
  population_size,
  numberof_days
)
}
\arguments{
\item{human}{human}

\item{IMild}{IMild}

\item{E}{E}

\item{ICase1}{ICase1}

\item{E_I}{list of individuals infected}

\item{p_E2_I}{p_E2_I}

\item{prob_hosp}{prob_hosp}

\item{population_size}{total population size}

\item{numberof_days}{number of days}
}
\description{
: NOTE function is not being used and maybe superceded by OJs work
Scheduler not tested yet
}
