% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbm_countData}
\alias{gbm_countData}
\title{Complete single-cell isoform raw count matrix.}
\format{
\subsection{\code{gbm_countData}}{

A dgCMatrix with 56,642 rows and 5,043 columns.
Row names are Ensembl transcript IDs and column names are cell IDs.
}
}
\source{
Package tutorial data.
}
\usage{
gbm_countData
}
\description{
Matrix containing raw isoform counts for tutorial use.
}
\keyword{datasets}
