% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbm_gtf}
\alias{gbm_gtf}
\title{Complete GTF of transcript structural annotations.}
\format{
\subsection{\code{gbm_gtf}}{

A GRanges object with 365,784 ranges and 7 metadata columns.
}
}
\source{
Package tutorial data.
}
\usage{
gbm_gtf
}
\description{
GRanges object of transcript structural annotations for tutorial use.
}
\keyword{datasets}
