% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{get_contact_matrix}
\alias{get_contact_matrix}
\title{Contact matrix at current timestep}
\usage{
get_contact_matrix(contact_matrix_set)
}
\arguments{
\item{contact_matrix_set}{model parameter for the contact matrices}
}
\value{
2D contact matrix for the current timestep
}
\description{
Currently only supports one contact matrix
}
