% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from SQUIRE model}
\usage{
get_parameters(pop, dt = 1, time_period = 365, ...)
}
\arguments{
\item{pop}{population. See \code{\link[squire]{get_population}}}

\item{dt}{the timestep (days) for the simulation}

\item{time_period}{the number timesteps to run the simulation for}

\item{...}{Other parameters for
\code{\link[squire]{parameters_explicit_SEEIR}}}
}
\value{
SQUIRE parameters
}
\description{
Get parameters from SQUIRE model
}
