% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_population}
\alias{get_population}
\title{Get population from squire model}
\usage{
get_population(country)
}
\arguments{
\item{country}{name of country}
}
\value{
population
}
\description{
rounds population sizes to discrete numbers
}
