% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{hospitilisation_flow_process}
\alias{hospitilisation_flow_process}
\title{Simulate hospitilisation flow}
\usage{
hospitilisation_flow_process(
  api,
  to_move,
  human,
  states,
  variables,
  parameters,
  events
)
}
\arguments{
\item{api}{simulation api}

\item{to_move}{individuals to move}

\item{human}{humans}

\item{states}{a list of all of the model states}

\item{variables}{a list of all of the model variables}

\item{parameters}{of the model}

\item{events}{a list of all of the model events}
}
\description{
Allocated individuals to hospital resources upon arrival
}
