% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{Individual_I_to_R_2}
\alias{Individual_I_to_R_2}
\title{Recovery_process -> I to R}
\usage{
Individual_I_to_R_2(I, R, human, immunity, age, location, pars = NULL)
}
\arguments{
\item{I}{infected}

\item{R}{recovered}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Recovery_process -> I to R
}
\examples{
Individual_I_to_R_2(I, R, human, immunity, age, location, pars)
}
