% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{Individual_R_to_S}
\alias{Individual_R_to_S}
\title{Recovered to susceptible -> R to S}
\usage{
Individual_R_to_S(S, R, human, immunity, age, location, pars = NULL)
}
\arguments{
\item{S}{susceptible}

\item{R}{recovered}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Recovered to susceptible -> R to S
}
\examples{
Individual_R_to_S(R, S, human, immunity, age, location, pars)
}
