% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{Individual_S_to_I}
\alias{Individual_S_to_I}
\title{Infection_process -> S to I}
\usage{
Individual_S_to_I(S, I, human, immunity, age, location, pars = NULL)
}
\arguments{
\item{S}{susceptible}

\item{I}{infected}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Infection_process -> S to I
}
\examples{
Individual_S_to_I(S, I, immunity, age, location, pars)
}
