% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{infected_scheduler}
\alias{infected_scheduler}
\title{IMild scheduler}
\usage{
infected_scheduler(human, E, numberof_days, event)
}
\arguments{
\item{human}{human}

\item{E}{human}

\item{numberof_days}{number of_days}

\item{event}{number of_daysan event}
}
\description{
IMild scheduler
}
