% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{infection}
\alias{infection}
\title{Function to find FOI and its probability}
\usage{
infection(IMild, ICase1, Icase2, dt, lambda, problambda)
}
\arguments{
\item{IMild}{Mild infections}

\item{ICase1}{First of the compartments for infections that will require hospitalisation}

\item{Icase2}{Second of the compartments for infections that will require hospitalisation}

\item{dt}{delta time}

\item{lambda}{FOI}

\item{problambda}{probability using lambda}

\item{api}{api}
}
\value{
list
}
\description{
Function to find FOI and its probability
}
\examples{
infection(IMild, ICase1, Icase2, dt, lambda, problambda)
}
