% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_process.R
\name{infection_process}
\alias{infection_process}
\title{Calculating the FOI}
\usage{
infection_process(
  human,
  IMild,
  ICase1,
  ICase2,
  S,
  E1,
  population_size,
  beta,
  m,
  dt
)
}
\arguments{
\item{human, }{human}

\item{IMild}{rest of the infections, which we consider to be mild and not
require hospitalisation}

\item{ICase1}{First of the compartments for infections that will require
hospitalisation}

\item{ICase2}{Second of the compartments for infections that will require
hospitalisation}

\item{S}{Susceptible}

\item{E1}{First of the latent infection compartments}

\item{population_size}{population_size}

\item{m}{m}

\item{dt}{dt}

\item{beta_set}{beta_set}
}
\description{
calculating the FOI and infection process
}
