% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{Init_check}
\alias{Init_check}
\title{Check and set up initial values}
\usage{
Init_check(init, population)
}
\arguments{
\item{init}{Data.frame of initial conditions. Default = NULL}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause the population to be sourced from \code{country}}
}
\value{
Checked initial values data.frame
}
\description{
Check and set up initial values
}
