% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{initialise_progression}
\alias{initialise_progression}
\title{Schedule progression of human disease at the start of the simulation}
\usage{
initialise_progression(event, human, from_state, duration)
}
\arguments{
\item{event}{the event to schedule}

\item{human}{the human handle}

\item{from_state}{the state this event applies to}

\item{duration}{the average time spent in this state}

\item{api}{the simulation api}
}
\description{
Schedules infection events using Erlang
}
