% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{Matrix_check}
\alias{Matrix_check}
\title{Check dimensions of inputs}
\usage{
Matrix_check(population, contact_matrix_set)
}
\arguments{
\item{population}{Population vector (for each age group). Default = NULL,
which will cause the population to be sourced from \code{country}}

\item{contact_matrix_set}{Contact matrices used in simulation. Default =
NULL, which will generate this based on the \code{country}.}
}
\value{
Null if checks pass
}
\description{
Check dimensions of inputs
}
