% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{Matrix_set}
\alias{Matrix_set}
\title{Process set of contact matrices ->  mixing matrices}
\usage{
Matrix_set(contact_matrix_set, population)
}
\arguments{
\item{contact_matrix_set}{Set of contact matrices}

\item{population}{Vector of populaion by age

#' @export}
}
\value{
Processed set of mixing matrices
}
\description{
Process set of contact matrices ->  mixing matrices
}
