% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{Matrix_set_explicit}
\alias{Matrix_set_explicit}
\title{Title}
\usage{
Matrix_set_explicit(contact_matrix_set, population)
}
\arguments{
\item{contact_matrix_set}{contact_matrix_set}

\item{population}{Vector of population by age}
}
\value{
Input matrix
}
\description{
Title
}
