% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{Pos_num}
\alias{Pos_num}
\title{Check argument is a single positive numeric}
\usage{
Pos_num(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{argument}

\item{name}{Name of argument}
}
\value{
Nothing if check pass
}
\description{
Check argument is a single positive numeric
}
