% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{Probabilities_of_states}
\alias{Probabilities_of_states}
\title{probabilities of the states}
\usage{
Probabilities_of_states(dt)
}
\arguments{
\item{dt}{time step}
}
\value{
states
}
\description{
probabilities of the states
}
\examples{
Probabilities_of_states(dt)
}
