% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrices.R
\name{Process_contact_matrix}
\alias{Process_contact_matrix}
\title{Process a contact matrix}
\usage{
Process_contact_matrix(contact_matrix, population)
}
\arguments{
\item{contact_matrix}{A contact matrix}

\item{population}{Vector of population by age}
}
\value{
Processed matrix
}
\description{
Process a contact matrix
}
