% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{SingleCellExperiment}
\alias{counts}
\alias{counts<-}
\alias{reducedDim}
\alias{reducedDim<-}
\alias{reducedDimNames}
\alias{reducedDimNames<-}
\alias{mainExpName}
\alias{mainExpName<-}
\alias{assay}
\alias{assay<-}
\alias{assayNames}
\alias{assays}
\alias{assays<-}
\alias{rowRanges}
\alias{rowRanges<-}
\alias{colData}
\alias{rowData}
\alias{rowData<-}
\alias{colData<-}
\alias{metadata}
\alias{metadata<-}
\alias{mcols}
\alias{mcols<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{S4Vectors}{\code{\link[S4Vectors:mcols]{mcols()}}, \code{\link[S4Vectors:mcols<-]{mcols<-()}}, \code{\link[S4Vectors:metadata]{metadata()}}, \code{\link[S4Vectors:metadata<-]{metadata<-()}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment:counts]{counts}}, \code{\link[SingleCellExperiment:counts<-]{counts<-}}, \code{\link[SingleCellExperiment:mainExpName]{mainExpName()}}, \code{\link[SingleCellExperiment:mainExpName<-]{mainExpName<-()}}, \code{\link[SingleCellExperiment:reducedDim]{reducedDim()}}, \code{\link[SingleCellExperiment:reducedDim<-]{reducedDim<-()}}, \code{\link[SingleCellExperiment:reducedDimNames]{reducedDimNames()}}, \code{\link[SingleCellExperiment:reducedDimNames<-]{reducedDimNames<-()}}, \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment()}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:assay]{assay()}}, \code{\link[SummarizedExperiment:assay<-]{assay<-()}}, \code{\link[SummarizedExperiment:assayNames]{assayNames()}}, \code{\link[SummarizedExperiment:assays]{assays()}}, \code{\link[SummarizedExperiment:assays<-]{assays<-()}}, \code{\link[SummarizedExperiment:colData]{colData()}}, \code{\link[SummarizedExperiment:colData<-]{colData<-()}}, \code{\link[SummarizedExperiment:rowData]{rowData()}}, \code{\link[SummarizedExperiment:rowData<-]{rowData<-()}}, \code{\link[SummarizedExperiment:rowRanges]{rowRanges}}, \code{\link[SummarizedExperiment:rowRanges<-]{rowRanges<-()}}}
}}

