% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{render_state_sizes2}
\alias{render_state_sizes2}
\title{Renders the sizes for S, I, R, I2 and D}
\usage{
render_state_sizes2(S, I, R, I2, D, human)
}
\arguments{
\item{S}{S}

\item{I}{I}

\item{R}{R}

\item{I2}{I2}

\item{D}{D}

\item{human}{human}
}
\description{
Renders the sizes for S, I, R, I2 and D
}
\examples{
render_state_sizes2(S, I, R, I2, D, human)
}
