% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{render_state_sizes3}
\alias{render_state_sizes3}
\title{Render sizes for SQUIRE states - note IMild is missing for now}
\usage{
render_state_sizes3(
  S,
  E1,
  E2,
  IMild,
  ICase1,
  ICase2,
  cum_hosp_inc,
  IOxGetLive1,
  IOxGetLive2,
  IOxNotGetLive1,
  IOxNotGetLive2,
  IOxGetDie1,
  IOxGetDie2,
  IOxNotGetDie1,
  IOxNotGetDie2,
  IMVGetLive1,
  IMVGetLive2,
  IMVNotGetLive1,
  IMVNotGetLive2,
  IMVGetDie1,
  IMVGetDie2,
  IMVNotGetDie1,
  IMVNotGetDie2,
  IRec1,
  IRec2,
  R,
  D,
  human
)
}
\arguments{
\item{S}{susceptible}

\item{E1}{First of the latent infection compartments}

\item{E2}{Second of the latent infection compartments}

\item{IMild}{rest of the infections, which we consider to be mild and not require hospitalisation}

\item{ICase1}{First of the compartments for infections that will require hospitalisation}

\item{ICase2}{Second of the compartments for infections that will require hospitalisation}

\item{cum_hosp_inc}{Those requiring hospitalisation}

\item{IOxGetLive1}{First of the compartments for infections that will require oxygen, get it, and who survive}

\item{IOxGetLive2}{Second of the compartments for infections that will require oxygen, get it, and who survive}

\item{IOxNotGetLive1}{First of the compartments for infections that will require oxygen, do NOT get it, and live}

\item{IOxNotGetLive2}{Second of the compartments for infections that will require oxygen, do NOT get it, and live}

\item{IOxGetDie1}{First of the compartments for infections that will require oxygen, get it, and die}

\item{IOxGetDie2}{Second of the compartments for infections that will require oxygen, get it, and die}

\item{IOxNotGetDie1}{First of the compartments for infections that will require oxygen, do NOT get it, and die}

\item{IOxNotGetDie2}{Second of the compartments for infections that will require oxygen, do NOT get it, and die}

\item{IMVGetLive1}{First of the compartments for infections that will require mechanical ventilation, get it, and who survive}

\item{IMVGetLive2}{Second of the compartments for infections that will require mechanical ventilation, get it, and who survive}

\item{IMVNotGetLive1}{First of the compartments for infections that will require mechanical ventilation, do NOT get it, and survive}

\item{IMVNotGetLive2}{Second of the compartments for infections that will require mechanical ventilation, do NOT get it, and survive}

\item{IMVGetDie1}{First of the compartments for infections that will require mechanical ventilation, get it, and die}

\item{IMVGetDie2}{Second of the compartments for infections that will require mechanical ventilation, get it, and die}

\item{IMVNotGetDie1}{First of the compartments for infections that will require mechanical ventilation, do NOT get it, and die}

\item{IMVNotGetDie2}{Second of the compartments for infections that will require mechanical ventilation, do NOT get it, and die}

\item{IRec1}{First of the compartments for those recovering from ICU}

\item{IRec2}{Second of the compartments for those recovering from ICU}

\item{R}{Recovered}

\item{D}{Dead}

\item{human}{human}
}
\description{
Render sizes for SQUIRE states - note IMild is missing for now
}
\examples{
render_state_sizes3(S, E1, E2, IMildICase1, ICase2, cum_hosp_inc, IOxGetLive1, IOxGetLive2, IOxNotGetLive1, IOxNotGetLive2, IOxGetDie1, IOxGetDie2, IOxNotGetDie1, IOxNotGetDie2, IMVGetLive1, IMVGetLive2, IMVNotGetLive1, IMVNotGetLive2, IMVGetDie1, IMVGetDie2, IMVNotGetDie1, IMVNotGetDie2, IRec1, IRec2, R, D, human)
}
