% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{Render_state_sizes_2}
\alias{Render_state_sizes_2}
\title{Renders the sizes for S, I, R}
\usage{
Render_state_sizes_2(S, I, R, human)
}
\arguments{
\item{S}{S}

\item{I}{I}

\item{R}{R}

\item{human}{human}
}
\description{
Renders the sizes for S, I, R
}
\examples{
Render_state_sizes_2(S, I, R, human)
}
