% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  country,
  timesteps = NULL,
  pop,
  parameters = NULL,
  contact_matrix_set,
  max_age = 100
)
}
\arguments{
\item{country}{country simulation is being run for}

\item{timesteps}{the number of timesteps to run the simulation for.}

\item{pop}{population. See \link[squire:get_population]{squire::get_population}}

\item{parameters}{parameters list.
See \link[squire:parameters_explicit_SEEIR]{squire::parameters_explicit_SEEIR}}

\item{contact_matrix_set}{contact_matrix_set}

\item{max_age}{maximum age defaults to 100}
}
\description{
The main entrypoint for the simulation. run_simulation puts together the
model components and runs the malaria simulation. This currently returns a
dataframe with the number of individuals in each state at each timestep
}
