% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_stochastic.R
\name{schedule_infections}
\alias{schedule_infections}
\title{Schedule infections}
\usage{
schedule_infections(api, events, inf_process)
}
\arguments{
\item{api}{api}

\item{events}{events}

\item{clinical_infections}{clinical infections}

\item{treated}{treated}

\item{infections}{infections}
}
\value{

}
\description{
Schedule infections
}
