% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRexplicitparameters.R
\name{SEIRexplicitparameters}
\alias{SEIRexplicitparameters}
\title{Calculating the FOI}
\usage{
SEIRexplicitparameters(
  human,
  IMild,
  ICase1,
  ICase2,
  cum_hosp_inc,
  i,
  N_age,
  lambda,
  problambda,
  beta,
  m,
  dt
)
}
\arguments{
\item{human, }{human}

\item{IMild}{rest of the infections, which we consider to be mild and not
require hospitalisation}

\item{ICase1}{First of the compartments for infections that will require
hospitalisation}

\item{ICase2}{Second of the compartments for infections that will require
hospitalisation}

\item{cum_hosp_inc}{cumulative hosptal incidences}

\item{i}{index}

\item{N_age}{number of points for a particular age}

\item{lambda}{FOI}

\item{problambda}{probability of}

\item{beta}{beta}

\item{m}{m}

\item{dt}{dt}
}
\description{
Calculating the FOI
}
\examples{
SEIRexplicitparameters(human, IMild, ICase1, ICase2, cum_hosp_inc, i, N_age,
lambda, problambda, beta, m, dt)
}
