% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRexplicitparameters.R
\name{SEIRexplicitparameters2}
\alias{SEIRexplicitparameters2}
\title{Calculting the FOI}
\usage{
SEIRexplicitparameters2(
  IMild,
  ICase1,
  ICase2,
  cum_hosp_inc,
  tt_matrix,
  mix_mat_set,
  tt_beta,
  beta_set,
  i,
  N_age,
  lambda,
  problambda
)
}
\arguments{
\item{IMild}{rest of the infections, which we consider to be mild and not require hospitalisation}

\item{ICase1}{First of the compartments for infections that will require hospitalisation}

\item{ICase2}{Second of the compartments for infections that will require hospitalisation}

\item{cum_hosp_inc}{cumulative hosptal incidences}

\item{tt_matrix}{time}

\item{mix_mat_set}{contact matrix}

\item{tt_beta}{time beta}

\item{beta_set}{beta_set}

\item{i}{index}

\item{N_age}{number of points for a particular age}

\item{lambda}{FOI}

\item{problambda}{probability of FOI}
}
\description{
Calculting the FOI
}
\examples{
SEIRexplicitparameters2(IMild, ICase1, ICase2, cum_hosp_inc,  tt_matrix, mix_mat_set, tt_beta, beta_set, i, N_age,lambda, problambda)
}
