% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{state_count_renderer_process}
\alias{state_count_renderer_process}
\title{Process to render number of individuals in specified states}
\usage{
state_count_renderer_process(individualname, statenames)
}
\arguments{
\item{individualname}{name of the individual of interest}

\item{statenames}{a list of state object's names}
}
\value{
rendered data
}
\description{
Process to render number of individuals in specified states
}
\examples{
state_count_renderer_process(individual, states)
}
