% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{validate_fixed_probability_state_change_process}
\alias{validate_fixed_probability_state_change_process}
\title{Function to check states}
\usage{
validate_fixed_probability_state_change_process(
  individual,
  state1,
  state2,
  prob
)
}
\arguments{
\item{individual}{individual}

\item{state1}{state1}

\item{state2}{individual}

\item{prob}{prob}

\item{api}{api}
}
\description{

}
\details{

}
\examples{
validate_fixed_probability_state_change_process(individual, state1, state2, prob)
}
