% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{validated_state_update}
\alias{validated_state_update}
\title{Check state update is valid}
\usage{
validated_state_update(api, i, state, index, population_size)
}
\arguments{
\item{api}{api}

\item{i}{human etc}

\item{state}{S, I, R etc}

\item{index}{recovered, susceptable, etc}

\item{population_size}{population size}
}
\description{
Check state update is valid
}
\examples{
validated_state_update(api, i, state, index, population_size)
}
