% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{create_age_variables}
\alias{create_age_variables}
\title{Create age variables}
\usage{
create_age_variables(pop)
}
\arguments{
\item{pop}{population list}
}
\value{
named list of individual::Variable
}
\description{
Create individual variables for continuous and discrete age
}
