% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{Create_discrete_age_variable}
\alias{Create_discrete_age_variable}
\title{Discrete age variable}
\usage{
Create_discrete_age_variable(pop)
}
\arguments{
\item{pop}{population list}
}
\value{
discrete age variable
}
\description{
Create a discrete age variable for each of the
length(pop$age_group) distinct age groups
}
