% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_individuals}
\alias{create_individuals}
\title{Define the human model returning more than 1 human}
\usage{
create_individuals(states, variables, events)
}
\arguments{
\item{states}{available states to assign}

\item{variables}{available variables to assign}

\item{events}{available events to assign}
}
\value{
list of humans
}
\description{
Define the human model returning more than 1 human
}
