% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{create_individuals}
\alias{create_individuals}
\title{Define model individuals}
\usage{
create_individuals(states, variables, events)
}
\arguments{
\item{states}{list of states}

\item{variables}{is a list of variables}

\item{events}{is a list of variables}
}
\description{
Create_individuals declares the individuals to simulate. It assigns the
relevant states and variables to each individual.
}
