% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\name{Create_individuals}
\alias{Create_individuals}
\title{Define model individuals}
\usage{
Create_individuals(states, variables, events, parameters)
}
\arguments{
\item{states}{available states to assign}

\item{variables}{available variables to assign}

\item{events}{available events to assign}

\item{parameters}{model parameters}
}
\description{
Create_individuals declares the individuals to simulate. It assigns the
relevant states and variables to each individual.
}
