% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create processes}
\usage{
create_processes(
  indivs,
  states,
  pstates,
  psq,
  population,
  beta,
  E_I,
  numberof_days
)
}
\arguments{
\item{indivs}{human, etc}

\item{states}{states}

\item{pstates}{probability of states}

\item{psq}{parameters list}

\item{population}{total population for all age groups}

\item{beta}{beta}

\item{E_I}{E_I}

\item{numberof_days}{number of days}
}
\value{
processes
}
\description{
Create processes for individual::simulate function
\itemize{
\item note that first 2 will be replaced with better functions
}
}
