% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create processes for simulation}
\usage{
create_processes(psq, pop, max_age = 100)
}
\arguments{
\item{psq}{model parameter}

\item{pop}{population information}

\item{max_age}{maximum age, defailt 100}
}
\value{
processes
}
\description{
Create processes for simulation
}
