% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create processes for simulation}
\usage{
create_processes(psq, pop)
}
\arguments{
\item{psq}{model parameter}

\item{pop}{population information}
}
\value{
processes
}
\description{
Create processes for simulation
}
