% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{create_processes}
\alias{create_processes}
\title{Create processes for simulation}
\usage{
create_processes(states, variables, events, individuals)
}
\arguments{
\item{states}{states for model}

\item{variables}{variables for model}

\item{events}{events for model}

\item{individuals}{individuals for model}
}
\value{
processes
}
\description{
Create processes for simulation
}
