% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{create_states}
\alias{create_states}
\title{Define model states}
\usage{
create_states(psq)
}
\arguments{
\item{psq}{the model parameters}
}
\value{
states
}
\description{
Create_states creates and initialises the human states for
the model
}
