% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSCE.R
\name{CreateSCE}
\alias{CreateSCE}
\title{Create a \emph{SingleCellExperiment} object}
\usage{
CreateSCE(
  countData,
  colData,
  rowData,
  active.gene.id = "gene_id",
  active.group.id = NULL,
  active.transcript.id = NULL,
  gtf = NULL,
  gtf.transcript.id = NULL,
  project = "Project",
  quiet = FALSE
)
}
\arguments{
\item{countData}{A numeric matrix containing raw single-cell isoform counts where row names are transcript IDs and column names are cell IDs.}

\item{colData}{A data frame of cell-level metadata. Row names must match the column names of \code{countData}.}

\item{rowData}{A data frame of isoform-level metadata. Row names must match the column names of \code{countData}.
Must include a column specifying gene IDs associated with each isoform as defined by the \code{active.gene.id} parameter.}

\item{active.gene.id}{Name of the column in \code{rowData} containing associated gene IDs associated with each isoform.
Defaults to \code{"gene_id"}.}

\item{active.group.id}{Name of the column in \code{colData} that defines cell grouping used in downstream functions. If \code{NULL}, no default grouping is assigned to the object metadata.}

\item{active.transcript.id}{Name of the column in \code{rowData} containing alternative transcript IDs (transcript names, ENST IDs, etc.) to use instead of row names of \code{countData}. Values in this column must be unique.}

\item{gtf}{A \code{GRanges} object containing structural annotations for all transcripts in \code{countData}.}

\item{gtf.transcript.id}{Name of metadata column in \code{gtf} that corresponds to the transcript IDs in \code{countData}.}

\item{project}{Project name to be stored in the object's \code{colData} slot.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Creates a \code{SingleCellExperiment} object from isoform counts, cell-level metadata, and isoform-level metadata.
}
