% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbm_rowData}
\alias{gbm_rowData}
\title{Complete transcript-level metadata of GBM sample.}
\format{
\subsection{\code{gbm_rowData}}{

A data frame with 56,642 rows and 7 columns:
\describe{
\item{\code{structural_category}}{Structural category of the transcript. One of full-splice_match", "incomplete-splice_match", "novel_in_catalog", "novel_not_in_catalog", "genic", "antisense", "fusion", "intergenic", "genic_intron". (SQANTI3}
\item{\code{gene_id}}{Associated Ensembl gene ID of the transcript.}
\item{\code{RTS_stage}}{Logical; \code{TRUE} if transcript contains one ore more RT switching artifacts. (SQANTI3)}
\item{\code{within_CAGE_peak}}{Logical; \code{TRUE} if transcript start site is within a CAGE peak. (SQANTI3)}
\item{\code{within_polyA_site}}{Logical; \code{TRUE} if transcript start site is within a polyA site. (SQANTI3)}
\item{\code{gene_name}}{Associated gene name of the transcript.}
\item{\code{transcript_name}}{Associated transcript name of the transcript.}
...
}
Row names are Ensembl transcript IDs.
}
}
\source{
Package tutorial data.
}
\usage{
gbm_rowData
}
\description{
Transcript-level metadata for tutorial use. Contains transcript quality control metrics generated by SQANTI3.
}
\keyword{datasets}
