% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from SQUIRE model}
\usage{
get_parameters(
  countryname,
  R0,
  dt,
  time_period,
  tt_contact_matrix,
  contact_matrix_set
)
}
\arguments{
\item{countryname}{name of country}

\item{R0}{R0}

\item{dt}{dt}

\item{time_period}{time period}

\item{tt_contact_matrix}{contact matrix time}

\item{contact_matrix_set}{contact_matrix_set}
}
\value{
SQUIRE parameters
}
\description{
Get parameters from SQUIRE model
}
