% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from SQUIRE model}
\usage{
get_parameters(
  countryname,
  R0,
  timestep,
  dt,
  time_period,
  tt_contact_matrix,
  newpopulation = NULL,
  numberof_days
)
}
\arguments{
\item{countryname}{name of country}

\item{R0}{R0}

\item{timestep}{tme step}

\item{dt}{dt}

\item{time_period}{time period}

\item{tt_contact_matrix}{contact matrix time}

\item{newpopulation}{population to use instead of actual one}

\item{numberof_days}{number of days for the run}
}
\value{
SQUIRE parameters
}
\description{
Get parameters from SQUIRE model
}
