% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from squire model}
\usage{
get_parameters(iso3c, time_period = 365, ...)
}
\arguments{
\item{iso3c}{three letter code for your country of interest}

\item{time_period}{for the simulation}

\item{...}{Other parameters for \link[squire:parameters_explicit_SEEIR]{squire::parameters_explicit_SEEIR}}
}
\value{
squire model parameters
}
\description{
Get parameters from squire model
}
