% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{get_population}
\alias{get_population}
\title{Get population from SQUIRE model}
\usage{
get_population(countryname)
}
\arguments{
\item{countryname}{name of country}
}
\description{
rounds population sizes to discrete numbers
}
