% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDiversity.R
\name{GetDiversity}
\alias{GetDiversity}
\title{Get summarized isoform diversity data}
\usage{
GetDiversity(
  object,
  genes,
  group.by = NULL,
  group.subset = NULL,
  method.use = "Tsallis",
  assay.use = "counts",
  diversity.cutoff = NULL,
  min.tx.cts = 1,
  order = NULL,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{genes}{A vector of one or more gene IDs.}

\item{group.by}{Name of \code{colData} variable to group cells. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.subset}{An optional vector specifying a subset of the elements in \code{group.by} to include.}

\item{method.use}{The diversity index to calculate.
Options include \code{"Tsallis"}, \code{"Shannon"}, \code{"NormalizedShannon"}, \code{"Renyi"}, \code{"NormalizedRenyi"}, \code{"GiniSimpson"}, or \code{InverseSimpson}.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{diversity.cutoff}{The cutoff of the diversity index used for monoform and polyform classification.
Default cutoffs are 0.243 for Tsallis, 0.500 for Shannon, 0 for normalized Shannon, 0.435 for Renyi, 0 for normalized Renyi, 0.348 for Gini-Simpson, and 1.533 for inverse Simpson.}

\item{min.tx.cts}{Minimum transcript counts required for a transcript to be included in the contingency table.}

\item{order}{Value specifying the order of entropy. Corresponds to \code{q} for Tsallis (default: 3) and \code{alpha} for Renyi (default: 2).}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{group}}{The cell group being queried.}
\item{\code{gene}}{The gene being queried}
\item{\code{gene.pct}}{Percentage of cells in \code{group} with expression of the gene.}
\item{\code{n.transcripts}}{Number of associated transcripts for the gene.}
\item{\code{transcript}}{The associated transcript.}
\item{\code{cts}}{Total counts of the transcript in \code{group}.}
\item{\code{prop}}{The transcript proportion in \code{group}.}
\item{\code{div}}{Isoform diversity of the gene in \code{group}.}
}
}
\description{
Retrieves summarized isoform diversity data of one or more genes.
}
