% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExpression.R
\name{GetExpression}
\alias{GetExpression}
\title{Get summarized isoform expression}
\usage{
GetExpression(
  object,
  transcripts,
  genes = NULL,
  group.by = NULL,
  group.subset = NULL,
  assay.use = "logcounts",
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{transcripts}{A vector containing one or more transcript IDs.}

\item{genes}{A vector containing one or more gene IDs. Will retrieve data for associated transcripts.}

\item{group.by}{Name of \code{colData} variable to group cells. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.subset}{An optional vector specifying a subset of the elements in \code{group.by} to include.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{group}}{The cell group being queried.}
\item{\code{gene}}{The (associated) gene being queried.}
\item{\code{transcript}}{The transcript being queried.}
\item{\code{transcript.pct}}{Percentage of cells in \code{group} with expression of the transcript.}
\item{\code{avgExpr}}{Average expression of the transcript across all cells in \code{group}.}
}
}
\description{
Retrieves summarized isoform expression data of one or more transcripts.
}
