% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUsage.R
\name{GetUsage}
\alias{GetUsage}
\title{Get summarized isoform usage data}
\usage{
GetUsage(
  object,
  genes,
  group.by = NULL,
  group.subset = NULL,
  assay.use = "counts",
  min.tx.cts = 1,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{genes}{A vector of one or more gene IDs.}

\item{group.by}{Name of \code{colData} variable to group cells. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.subset}{An optional vector specifying a subset of the elements in \code{group.by} to include.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{min.tx.cts}{Minimum transcript counts required for a transcript to be included in proportion calculations.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{group}}{The cell group being queried.}
\item{\code{gene}}{The gene being queried.}
\item{\code{gene.pct}}{Percentage of cells in \code{group} with expression of the gene.}
\item{\code{transcript}}{The associated transcript.}
\item{\code{cts}}{Total counts of the transcript across cells in \code{group}.}
\item{\code{prop}}{Transcript proportion.}
}
}
\description{
Retrieves summarized isoform usage data of one or more genes.
}
