% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{individual_I_to_R_I2_to_D}
\alias{individual_I_to_R_I2_to_D}
\title{Recovery_process and death process -> I to R and I2 to D}
\usage{
individual_I_to_R_I2_to_D(
  I,
  R,
  I2,
  D,
  human,
  immunity,
  age,
  location,
  pars = NULL
)
}
\arguments{
\item{I}{infected}

\item{R}{recovered}

\item{I2}{severely infected}

\item{D}{dead}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Recovery_process and death process -> I to R and I2 to D
}
\examples{
individual_I_to_R_I2_to_D(I, R, I2, D, human, immunity, age, location, pars)
}
