% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{individual_S_to_I_and_I2}
\alias{individual_S_to_I_and_I2}
\title{Infection_process -> S to I and I2}
\usage{
individual_S_to_I_and_I2(S, I, I2, human, immunity, age, location, pars = NULL)
}
\arguments{
\item{S}{susceptible}

\item{I}{infected}

\item{I2}{severelyinfected}

\item{human}{human}

\item{immunity}{immunity}

\item{age}{age}

\item{location}{location}

\item{pars}{parameter list}
}
\description{
Infection_process -> S to I and I2
}
\examples{
individual_S_to_I_and_I2(S, I, I2, human, immunity, age, location, pars)
}
