% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_process.R
\name{infection_process}
\alias{infection_process}
\title{Calculating the FOI}
\usage{
infection_process(
  human,
  S,
  E,
  population_size,
  beta,
  m,
  dt,
  IMild,
  ICase1,
  ICase2
)
}
\arguments{
\item{human, }{human}

\item{S}{Susceptible}

\item{E}{Infection compartments}

\item{population_size}{population_size}

\item{beta}{beta}

\item{m}{m}

\item{dt}{dt}

\item{IMild}{rest of the infections, which we consider to be mild and not
require hospitalisation}

\item{ICase1}{First of the compartments for infections that will require
hospitalisation}

\item{ICase2}{Second of the compartments for infections that will require
hospitalisation}
}
\description{
calculating the FOI and infection process
}
\section{NOTE function is not being used and maybe superceded by OJs work}{
###NOTE: # E_I not being used in E_IMild_ICase1_ICase2 for time being
}

