% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_process.R
\name{infection_process}
\alias{infection_process}
\title{Calculating the FOI}
\usage{
infection_process(human, states, discrete_age, exposure)
}
\arguments{
\item{human}{the human handle}

\item{states}{a list of states in the model}

\item{discrete_age}{variable}

\item{exposure}{event for covid exposure}
}
\description{
calculating the FOI and infection process
}
