% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_process.R
\name{infection_process}
\alias{infection_process}
\title{Calculating the FOI}
\usage{
infection_process(individuals, states, variables, events)
}
\arguments{
\item{individuals}{a list of individuals in the model}

\item{states}{a list of states in the model}

\item{variables}{list of variables in the model}

\item{events}{a list of events in the model}
}
\description{
calculating the FOI and infection process
}
