% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeCounts.R
\name{NormalizeCounts}
\alias{NormalizeCounts}
\title{Normalize raw isoform counts}
\usage{
NormalizeCounts(
  object,
  method.use = "LogNormalize",
  scale.factor = 10000,
  gtf = NULL,
  gtf.transcript.id = NULL,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{method.use}{Normalization method. Options include \code{"LogNormalize"}, \code{"TPM"}, and \code{"FT"}. A new assay will be created (or overwritten) with name "logcounts", "tpmcounts", or "ftcounts", respectively.}

\item{scale.factor}{Scale factor for log and Freeman-Tukey normalization methods.}

\item{gtf}{A \code{GRanges} object containing transcript structural annotations used for TPM normalization.}

\item{gtf.transcript.id}{Metadata column name in \code{gtf} containing transcript IDs.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
Returns the object with a new assay containing normalized counts.
}
\description{
Normalizes raw isoform counts.
}
