% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCellQC.R
\name{PlotCellQC}
\alias{PlotCellQC}
\title{Visualize single-cell QC metrics}
\usage{
PlotCellQC(
  object,
  group.by = "project",
  colors = NULL,
  pt.size = 0.2,
  pt.alpha = 1,
  text.size = 12,
  show.legend = TRUE,
  combine = TRUE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{group.by}{One or more \code{colData} column names used to define cell groups.}

\item{colors}{A vector of colors to use for the plot.}

\item{pt.size}{Point size.}

\item{pt.alpha}{Point alpha.}

\item{text.size}{Text size.}

\item{show.legend}{Logical; if \code{TRUE}, the legend will be shown.}

\item{combine}{Logical; if \code{TRUE}, combines plots using \code{patchwork}.}
}
\value{
A patchwork object when \code{combine = TRUE}; otherwise a named list of ggplot objects.
}
\description{
Plots \code{nTranscript}, \code{nGene}, and \code{nCount} QC metrics by cell group.
}
