% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiversity.R
\name{PlotDiversity}
\alias{PlotDiversity}
\title{Visualize isoform diversity}
\usage{
PlotDiversity(
  object,
  genes,
  group.by = NULL,
  group.subset = NULL,
  group.order = NULL,
  plot.type = "lollipop",
  entropy.use = "Tsallis",
  assay.use = "counts",
  entropy.thresh = NULL,
  min.tx.cts = 1,
  order = NULL,
  colors = NULL,
  text.size = 12,
  nrow = 1,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{genes}{A vector of one or more gene IDs.}

\item{group.by}{Name of \code{colData} variable to group cells. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.subset}{An optional vector specifying a subset of the elements in \code{group.by} to include.}

\item{group.order}{An optional vector specifying the order of group elements.}

\item{plot.type}{Which type of plot to generate. Options include \code{"lollipop"}, \code{"density"}, and \code{"pcoord"}.}

\item{entropy.use}{The diversity index to calculate.
Options include \code{"Tsallis"}, \code{"Shannon"}, \code{"NormalizedShannon"}, \code{"Renyi"}, \code{"NormalizedRenyi"}, \code{"GiniSimpson"}, or \code{InverseSimpson}.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{entropy.thresh}{The threshold of the diversity index used for monoform and polyform classification.
Default cutoffs are 0.243 for Tsallis, 0.500 for Shannon, 0.500 / log(2) (~0.721) for normalized Shannon, 0.435 for Renyi, 0.435 / log(2) (~0.628) for normalized Renyi, 0.348 for Gini-Simpson, and 1.533 for inverse Simpson.}

\item{min.tx.cts}{Minimum transcript counts required for a transcript to be included in the contingency table.}

\item{order}{Value specifying the order of entropy. Corresponds to \code{q} for Tsallis (default: 3) and \code{alpha} for Renyi (default: 2).}

\item{colors}{A vector of colors to use for the plot.}

\item{text.size}{Text size.}

\item{nrow}{Number of facet rows.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A ggplot object.
}
\description{
Generates plots showing the isoform diversity of one or more genes.
}
