% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotExpression.R
\name{PlotExpression}
\alias{PlotExpression}
\title{Visualize isoform expression}
\usage{
PlotExpression(
  object,
  transcripts,
  group.by = NULL,
  group.subset = NULL,
  group.order = NULL,
  plot.type = "violin",
  colors = NULL,
  colors.heatmap = NULL,
  dim.use = NULL,
  assay.use = "logcounts",
  scale.heatmap = TRUE,
  pt.size = 0.1,
  pt.alpha = 1,
  text.size = 12,
  label = TRUE,
  show.gene = TRUE,
  nrow = 1,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{transcripts}{A vector of one or more transcript IDs.}

\item{group.by}{Name of \code{colData} variable to group cells. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.subset}{An optional vector specifying a subset of the elements in \code{group.by} to include.}

\item{group.order}{An optional vector specifying the order of group elements.}

\item{plot.type}{Which type of plot to generate. Options include \code{"violin"}, \code{"reducedDim"}, and \code{"heatmap"}.}

\item{colors}{A vector of colors to use for the plot.}

\item{colors.heatmap}{A vector of colors for the heatmap color bar.}

\item{dim.use}{Name of dimensions to use in the \code{reducedDim} slot of the object. This parameter is for the \code{reducedDim} plot type.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{scale.heatmap}{Logical; if \code{TRUE}, standardizes and clips expression values for the heatmap.}

\item{pt.size}{Point size for the \code{violin} and \code{reducedDim} plots.}

\item{pt.alpha}{Point alpha for the \code{violin} and `reducedDim`` plots.}

\item{text.size}{Text size.}

\item{label}{Logical; if \code{TRUE}, adds cell group labels to the \code{reducedDim} and \code{heatmap} plots.}

\item{show.gene}{Logical; if \code{TRUE}, shows the active gene ID alongside transcript names.}

\item{nrow}{Number of facet rows.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A ggplot object.
}
\description{
Generates plots showing the isoform expression of one or more transcripts.
}
