% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUsage.R
\name{PlotUsage}
\alias{PlotUsage}
\title{Visualize isoform usage across cell groups}
\usage{
PlotUsage(
  object,
  gene,
  group.by = NULL,
  group.subset = NULL,
  group.order = NULL,
  plot.type = "stackedbar",
  colors = NULL,
  show.prop = FALSE,
  nrow = 1,
  assay.use = "counts",
  min.tx.cts = 1,
  min.tx.prop = 0,
  text.size = 12,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{gene}{A gene ID.}

\item{group.by}{Name of \code{colData} variable to group cells. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.subset}{An optional vector specifying a subset of the elements in \code{group.by} to include.}

\item{group.order}{An optional vector specifying the order of group elements.}

\item{plot.type}{Which type of plot to generate. Options include \code{"stackedbar"}, \code{"bar"}, \code{"pie"}, and \code{"heatmap"}.}

\item{colors}{A vector of colors to use for the plot.}

\item{show.prop}{Logical; if \code{TRUE}, proportion labels will be included on the plot.}

\item{nrow}{Number of facet rows.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{min.tx.cts}{Minimum transcript counts in at least 1 group required for a transcript to be included in proportion calculations.}

\item{min.tx.prop}{Minimum transcript proportion required for a transcript to be plotted. Note: This parameter is intended for use with the \code{"heatmap"} plot type.}

\item{text.size}{Text size.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A ggplot object.
}
\description{
Generates plots showing the isoform usage of a gene.
}
