% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{probabilities_of_states}
\alias{probabilities_of_states}
\title{probabilities of the states}
\usage{
probabilities_of_states(dt, psq)
}
\arguments{
\item{dt}{time step}

\item{psq}{parameters}
}
\value{
states
}
\description{
probabilities of the states
}
