% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{Render_state_sizes4}
\alias{Render_state_sizes4}
\title{Renders the sizes for S, E1, E2, I}
\usage{
Render_state_sizes4(S, E1, E2, I, human)
}
\arguments{
\item{S}{S}

\item{E1}{E1}

\item{E2}{E2}

\item{I}{I}

\item{human}{human}
}
\description{
Renders the sizes for S, E1, E2, I
}
\examples{
Render_state_sizes4(S, E1, E2, I, human)
}
