% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(timesteps, pop, parameters = NULL, ...)
}
\arguments{
\item{timesteps}{the number of timesteps to run the simulation for}

\item{pop}{population. See \code{\link{squire::get_population}}}

\item{...}{Other parameters for \code{\link{squire::parameters_explicit_SEEIR}}}
}
\description{
The main entrypoint for the simulation. run_simulation puts together the
model components and runs the malaria simulation. This currently returns a
dataframe with the number of individuals in each state at each timestep
}
