% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\usage{
run_simulation(
  countryname,
  R0,
  timestep,
  dt,
  time_period,
  tt_contact_matrix,
  newpopulation = NULL,
  numberof_days
)
}
\arguments{
\item{countryname}{country name}

\item{R0}{R0}

\item{timestep}{time step}

\item{dt}{dt}

\item{time_period}{time period}

\item{tt_contact_matrix}{tt contact matrix}

\item{newpopulation}{set this to NULL to use pop$n}

\item{numberof_days}{number of days to run for}
}
\value{
data frame
}
\description{
Run the simulation for a given country
}
