% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDEI.R
\name{RunDEI}
\alias{RunDEI}
\title{Run differential isoform expression analysis}
\usage{
RunDEI(
  object,
  group.by = NULL,
  group.1 = NULL,
  group.2 = NULL,
  assay.use = "logcounts",
  min.pct = 0.01,
  only.pos = FALSE,
  transcripts = NULL,
  p.adj = "bonferroni",
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{group.by}{One or more \code{colData} column names used to define cell groups. If \code{NULL}, \code{metadata(object)$active.group.id} is used.}

\item{group.1}{Group label(s) for the first side of the comparison. If \code{NULL}, each group is compared against all others.}

\item{group.2}{Optional group label(s) for the second side of the comparison. If \code{NULL}, \code{group.1} is compared against all other cells.}

\item{assay.use}{Assay name to use. The default and recommended assay is \code{"logcounts"}.}

\item{min.pct}{Minimum fraction of cells in each group where the transcript must be detected.}

\item{only.pos}{Logical; if \code{TRUE}, only transcripts with positive fold change will be reported.}

\item{transcripts}{Optional vector of active transcript IDs to test.}

\item{p.adj}{P-value adjustment method. Must be one of \code{stats::p.adjust.methods}.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{group.1} & \code{group.2}}{The two cell groups being compared.}
\item{\code{gene}}{The gene associated with the transcript being tested.}
\item{\code{transcript}}{The transcript being tested.}
\item{\code{pct.1}}{Percentage of cells in \code{group.1} with expression of the transcript.}
\item{\code{pct.2}}{Percentage of cells in \code{group.2} with expression of the transcript.}
\item{\code{avgExpr.1}}{Average expression of the transcript in \code{group.1}.}
\item{\code{avgExpr.2}}{Average expression of the transcript in \code{group.2}.}
\item{\code{log2FC}}{The log2 fold change in transcript expression between the two groups (\code{group.1} - \code{group.2}).}
\item{\code{pval}}{P-value from the Wilcoxon rank-sum test.}
\item{\code{padj}}{Adjusted p-value, calculated separately for each group comparison.}
}
}
\description{
Tests transcripts for differential isoform expression between cell groups.
}
