% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDEI.R
\name{RunDEI}
\alias{RunDEI}
\title{Run differential isoform expression analysis}
\usage{
RunDEI(
  object,
  group.by = NULL,
  group.1 = NULL,
  group.2 = NULL,
  assay.use = "logcounts",
  min.pct = 0.01,
  only.pos = FALSE,
  transcripts = NULL,
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{group.by}{Name of \code{colData} variable to group cells for comparisons. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.1}{First group in the comparison.}

\item{group.2}{Second group in the comparison.}

\item{assay.use}{Which \code{assay} (counts) to use. The default and recommended option is \code{"logcounts"}.}

\item{min.pct}{Minimum percentage of cells in which a transcript must be expressed in both groups for it to be adjusted for multiple testing and reported in results.}

\item{only.pos}{Logical; if \code{TRUE}, only transcripts with positive fold change will be reported.}

\item{transcripts}{A vector of transcript IDs to test.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A data frame with the following columns:
\describe{
\item{\code{group.1} & \code{group.1}}{The two cell groups being compared.}
\item{\code{gene}}{The gene associated with the transcript being tested.}
\item{\code{transcript}}{The transcript being tested.}
\item{\code{pct.1}}{Average expression of the transcript across all cells in \code{group.1}.}
\item{\code{pct.2}}{Average expression of the transcript across all cells in \code{group.2}.}
\item{\code{log2FC}}{The log2 fold change in transcript expression between the two groups (\code{group.1} - \code{group.2}).}
\item{\code{pval}}{P-value from the the Wilcoxon rank-sum test.}
\item{\code{padj}}{Adjusted p-value, calculated separately for each group comparison (default: Bonferroni).}
}
}
\description{
Runs differential isoform expression analysis, which compares the mean isoform expression across two groups of cells using Wilcoxon-based testing.
}
