% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDIU.R
\name{RunDIU}
\alias{RunDIU}
\title{Run differential isoform usage analysis}
\usage{
RunDIU(
  object,
  group.by = NULL,
  group.1 = NULL,
  group.2 = NULL,
  assay.use = "counts",
  method.use = "Chisq",
  min.gene.pct = 0.05,
  min.gene.cts = 15,
  min.tx.cts = 1,
  genes = NULL,
  only.valid = FALSE,
  simulate.p = FALSE,
  p.adj = "BH",
  quiet = FALSE
)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{group.by}{Name of \code{colData} variable to group cells for comparisons. If \code{NULL}, \code{metadata(object)$active.group.id} will be used.}

\item{group.1}{First group in the comparison.}

\item{group.2}{Second group in the comparison.}

\item{assay.use}{Which \code{assay} (counts) to use.}

\item{method.use}{Statistical test, either "Chisq" for Chi-squared test or "Fisher" for Fisher's exact test.}

\item{min.gene.pct}{Minimum percentage of cells in which a gene must be expressed in both groups for it to be tested.}

\item{min.gene.cts}{Minimum total transcript counts in which a gene must have in both groups for it be tested.}

\item{min.tx.cts}{Minimum transcript counts required for a transcript to be included in contingency tables.}

\item{genes}{Vector of genes to test. Note: genes will still be subject to filtering.}

\item{only.valid}{Logical; if \code{TRUE}, only tests with valid Chi-square approximations will be reported and corrected for multiple testing.}

\item{simulate.p}{Logical; if \code{TRUE}, p-values are computed using a Monte Carlo simulation. Note: Fisher's exact test always uses simulation.}

\item{p.adj}{Method for p-value adjustment. Options are \code{"BH"} (Benjamini-Hochberg) or \code{"Bonferroni"}.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
A list containing two data frames:

\describe{
\item{\verb{$data}}{
A data frame of summarized data with columns:
\describe{
\item{\code{group.1} & \code{group.2}}{The two cell groups being compared.}
\item{\code{gene}}{The gene being tested.}
\item{\code{gene.pct.1}}{Percentage of cells in \code{group.1} with expression of the gene.}
\item{\code{gene.pct.2}}{Percentage of cells in \code{group.2} with expression of the gene.}
\item{\code{transcript}}{The associated transcript.}
\item{\code{cts.1}}{Total counts of the transcript across all cells in \code{group.1}.}
\item{\code{cts.2}}{Total counts of the transcript across all cells in \code{group.2}.}
\item{\code{prop.1}}{Transcript proportion for \code{group.1}.}
\item{\code{prop.2}}{Transcript proportion for \code{group.2}.}
\item{\code{delta}}{The difference in transcript proportions between groups (\code{group.1} - \code{group.2}).}
}
}

\item{\verb{$stats}}{
A data frame containing statistical results with columns:
\describe{
\item{\code{group.1} & \code{group.2}}{The two cell groups being compared.}
\item{\code{gene}}{The gene being tested.}
\item{\code{max.delta}}{The largest absolute difference in transcript proportions between \code{group.1} and \code{group.2}.}
\item{\code{transcript}}{The transcript associated with \code{max.delta}.}
\item{\code{pval}}{P-value from the statistical test specified in \code{method.use} (default: Chi-square test).}
\item{\code{effect.size}}{Effect size of the test, measured as Cramer's V.}
\item{\code{approx}}{Indicates whether the Chi-square approximation is valid ("valid") or potentially unreliable ("warning"), based on whether at least 80\% of transcript counts of the contingency table exceed 5.}
}
}
}
}
\description{
Runs differential isoform usage analysis.
For each comparison, genes are first filtered according to coverage, counts are aggregated across groups, then the Chi-square or Fisher's exact test is applied.
}
