% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRexplicitparameters.R
\name{SEIRexplicitparameters}
\alias{SEIRexplicitparameters}
\title{Setting the parameters for Explicit SEIR}
\usage{
SEIRexplicitparameters(
  S,
  E1,
  E2,
  IMild,
  ICase1,
  ICase2,
  cum_hosp_inc,
  cum_ICU_inc,
  IOxGetLive1,
  IOxGetLive2,
  IOxGetDie1,
  IOxGetDie2,
  IOxNotGetLive1,
  IOxNotGetLive2,
  IOxNotGetDie1,
  IOxNotGetDie2,
  IMVGetLive1,
  IMVGetLive2,
  IMVGetDie1,
  IMVGetDie2,
  IMVNotGetLive1,
  IMVNotGetLive2,
  IMVNotGetDie1,
  IMVNotGetDie2,
  IRec1,
  IRec2,
  R,
  D,
  D_get,
  D_not_get,
  i,
  j
)
}
\value{

}
\description{
Setting the parameters for Explicit SEIR
}
