% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetGenes.R
\name{SetGenes}
\alias{SetGenes}
\title{Set the active gene ID.}
\usage{
SetGenes(object, id, quiet = FALSE)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{id}{Name of \code{rowData} column containing gene IDs associated with each transcript.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
Returns object with updated \code{active.gene.id} stored in \code{metadata(object)}.
}
\description{
Updates the active gene ID (gene symbols, ENSG IDs, etc.) for downstream functions.
}
