% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetGroups.R
\name{SetGroups}
\alias{SetGroups}
\title{Set active cell groups}
\usage{
SetGroups(object, id)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{id}{Name of the \code{colData} column containing cell groups.
Character and factor columns are supported.}
}
\value{
The object with \code{metadata(object)$active.group.id} updated.
}
\description{
Selects the \code{colData} column used as the default cell grouping in downstream functions.
}
