% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetGroups.R
\name{SetGroups}
\alias{SetGroups}
\title{Set the active cell group}
\usage{
SetGroups(object, id)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{id}{Name of \code{colData} column containing the desired cell groups (cell types, cell states, sample ID, etc.).}
}
\value{
Returns object with updated \code{active.group.id} stored in \code{metadata(object)}.
}
\description{
Sets or updates the active cell group for downstream functions.
}
