% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetTranscripts.R
\name{SetTranscripts}
\alias{SetTranscripts}
\title{Set active transcript IDs}
\usage{
SetTranscripts(object, id)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{id}{Name of the \code{rowData} column containing unique transcript IDs. Use \code{""} to report row names.}
}
\value{
The object with \code{metadata(object)$active.transcript.id} updated.
}
\description{
Selects the \code{rowData} column used as transcript IDs in downstream functions.
}
