% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetTranscripts.R
\name{SetTranscripts}
\alias{SetTranscripts}
\title{Set the active transcript ID}
\usage{
SetTranscripts(object, id)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{id}{Name of \code{rowData} column containing unique transcript IDs.
Set to \code{""} to use the transcript IDs (row names) of original count matrix.}
}
\value{
Returns object with updated \code{active.transcript.id} stored in \code{metadata(object)}.
}
\description{
Update the active transcript ID for downstream functions.
}
