% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{state_count_renderer_process}
\alias{state_count_renderer_process}
\title{Process to render number of individuals in specified states}
\usage{
state_count_renderer_process(individual, states)
}
\arguments{
\item{individual}{the individual of interest}

\item{states}{a list of state objects}
}
\description{
Process to render number of individuals in specified states
}
\examples{
state_count_renderer_process(individual, states)
}
