% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetCells.R
\name{SubsetCells}
\alias{SubsetCells}
\title{Subset by cells}
\usage{
SubsetCells(object, subset = NULL, cells = NULL, invert = FALSE, quiet = FALSE)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{subset}{A logical expression evaluated within the \code{colData} of the object.
Rows (cells) for which the expression evaluates to \code{TRUE} will be retained. Supports column names from \code{colData} and standard logical operators.}

\item{cells}{A vector of cell IDs for filtering.}

\item{invert}{Logical; if \code{TRUE}, inverts the selection to remove the cells instead.}

\item{quiet}{Logical; if \code{TRUE}, suppresses messages.}
}
\value{
Returns the object after cell selection.
}
\description{
Subsets a \code{SingleCellExperiment} object by cells.
}
