% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRmodels.R
\name{validated_state_update}
\alias{validated_state_update}
\title{Check state update is valid}
\usage{
validated_state_update(i, state, ix, population_size)
}
\arguments{
\item{i}{human etc}

\item{state}{S, I, R, etc}

\item{ix}{index}

\item{population_size}{population size}
}
\description{
Check state update is valid
}
